Update OPENAQAPI key assignment in mlpred.py
In AWS, the OpenAQ API key should be available as an environment variable "OPENAQ_API_KEY". However, this seems not to be working properly. Until I can assess why, I am going to hard-code in the API key for testing purposes, leaving the "correct" access method commented out for now.